python bindings: define common types/methods to unify string handling between python2 and 3
It does not properly solve all issues. Basically: python3 now treats
all strings as unicode strings, but the ctypes API conversion for
char* (ctypes.c_char_p) is mapped to byte strings. An explicit
conversion has to occur, but we have to specify the appropriate
encoding, which is hard to guess.