Add automated testing via Travis CI and flake8
Travis CI can be used to do automated testing on all pull requests before they are reviewed and merged. This service is free to open source projects like this one. For this to work, @oaubert would need to turn on the switch for this repo at https://www.travis-ci.com/oaubert (login via GitHub credentials)