Another tkvlc.py update for macOS.
The menu shortcut keys now work on macOS, with a little cleaner workaround for the missing and/or failing modifier key binding. 

Also, the Play and Mute menu and button labels now toggle between Play and Pause respectively Mute and Unmute.

The timer class has been replaced by a one-line call to Tk's after method.

The Quit menu, the Q- and W-shortcut keys and the error dialog are all now working properly.

Unresolved remains the overwriting of the buttons and sliders by the video.  The buttons do work when clicking near the bottom of the video.

Un/mute is working, but may be unreliable (a VLC issue, see the libvlc_audio_set_mute.__doc__).

Like before, only tested on macOS 10.13.6 with Tk 8.6.9 and vlc.py 3.0.6109 using Python 3.7.4 64-bit. Some screen shots follow.