generator: fix first function parse
The "#define" regex does not seem to match empty value #define.
For example, the "# define LIBVLC_API" line is not considered as define
sequence. One of the consequence is the first function declaration
detected is considered as part of a multiline block, and thus ignored.

Accepting "#define XXX" sequences as well as "#define XXX yyy" seems to
fix the issue (first function declared in header missing).