Improved tkvlc.py example (#233)
This is tkvlc.py modified to make dragging over the buttons panel work and two other options.  

The original code to detach, un-anchor the buttons panel by press-move-release the left mouse button never worked properly.  In addition, if the buttons panel is un-anchored, it is stuck somewhere in the middle of the screen, can't be moved and can't be re-anchored to the video panel.

A new menu option Buttons Up/Down and a new button Up/Down have been added to provide a better way to anchor the buttons panel under the video panel (aka Down) or in the top-left corner of the screen (aka Up).

Also, dragging does work, but is disabled by default.  Use command line option -dragging to enable that feature.