Double events.
This version of tkvlc.py includes a workaround for another issue: double Tk events for accelerator keys defined with the Shift modifier.

When using the Cmd-Shift-+ key to Zoom In, two events are generated resulting in twice the zoom factor.  Similarly for the Cmd-Shift-> and Cmd-Shift-< keys to speed to video up or down.  Using menu selection works fine.

The problem occurs only with recent Python releases on (recent) macOS versions and both on Intel and Apple Silicon.  An issue is being prepared for submission at the Python GitHub site.