Instance: add a set_logger method
Adds a way to link libVLC log system to a python logging.Logger object.

The logger and log callback is attached to the libVLC Instance object to avoid
the python garbage collector to destroy them too soon.

NOTE: as libVLC log messages are based on a printf + va_list format, we
need to use the standard C `vsnprintf` function with a max fixed size output.