Partially fix regression in [OUT] parameter handling
Part of the problem was that `Func.docs_in_sphinx_format`
is called after `Par.flags` during generation, when the former
method fills the `out` member of `Func`, on which the output
of `Par.flags` depends.

It resulted in out parameters being incorrectly flagged as in
parameters (because it is the default).

The concerned functions are:

- `libvlc_log_get_context`
- `libvlc_log_get_object`.